# Support configurable quote character parameter 