--- a/mint/Dockerfile.dev
+++ b/mint/Dockerfile.dev
@@ -74,6 +74,9 @@ RUN build/worm/install.sh
 COPY build/healthcheck /mint/build/healthcheck
 RUN build/healthcheck/install.sh
 
+COPY build/s3select /mint/build/s3select
+RUN build/s3select/install.sh
+
 COPY remove-packages.list /mint
 COPY postinstall.sh /mint
 RUN /mint/postinstall.sh