## Description
Quote character was hardcoded and only `"` is accepted. This PR makes quote character configurable and can accept any single utf8 character

QuoteEscapeCharacter is still not configurable and will always get the value QuoteCharacter

## Motivation and Context
Add support of quote character configuration in CSV Input & Output character

## How to test this PR?
Test the minio-py example: select_object_content.py

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
- [ ] Fixes a regression (If yes, please add `commit-id` or `PR #` here)
- [ ] Documentation needed
- [ ] Unit tests needed
- [x] Functional tests needed (If yes, add [mint](https://github.com/minio/mint) PR # here: )