@vadmeste 2020-02-17
@klauspost this is the current benchmark regression:

```
benchmark                                            old ns/op     new ns/op     delta
BenchmarkRead-4                                      2622          3352          +27.84%
BenchmarkReadWithFieldsPerRecord-4                   2576          3643          +41.42%
BenchmarkReadWithoutFieldsPerRecord-4                2559          3382          +32.16%
BenchmarkReadLargeFields-4                           4666          5539          +18.71%
BenchmarkReadReuseRecord-4                           1430          2268          +58.60%
BenchmarkReadReuseRecordWithFieldsPerRecord-4        1420          2293          +61.48%
BenchmarkReadReuseRecordWithoutFieldsPerRecord-4     1455          2208          +51.75%
BenchmarkReadReuseRecordLargeFields-4                3399          4271          +25.65%

benchmark                                            old allocs     new allocs     delta
BenchmarkRead-4                                      16             26             +62.50%
BenchmarkReadWithFieldsPerRecord-4                   16             26             +62.50%
BenchmarkReadWithoutFieldsPerRecord-4                16             26             +62.50%
BenchmarkReadLargeFields-4                           24             36             +50.00%
BenchmarkReadReuseRecord-4                           6              16             +166.67%
BenchmarkReadReuseRecordWithFieldsPerRecord-4        6              16             +166.67%
BenchmarkReadReuseRecordWithoutFieldsPerRecord-4     6              16             +166.67%
BenchmarkReadReuseRecordLargeFields-4                12             24             +100.00%

benchmark                                            old bytes     new bytes     delta
BenchmarkRead-4                                      664           672           +1.20%
BenchmarkReadWithFieldsPerRecord-4                   664           672           +1.20%
BenchmarkReadWithoutFieldsPerRecord-4                664           672           +1.20%
BenchmarkReadLargeFields-4                           3936          3948          +0.30%
BenchmarkReadReuseRecord-4                           24            32            +33.33%
BenchmarkReadReuseRecordWithFieldsPerRecord-4        24            32            +33.33%
BenchmarkReadReuseRecordWithoutFieldsPerRecord-4     24            32            +33.33%
BenchmarkReadReuseRecordLargeFields-4                2976          2988          +0.40%
```