@vadmeste 2020-02-17
@harshavardhana actually some variables can be cached between CSV Read() calls, this is the optimization result:

```
benchmark                                            old ns/op     new ns/op     delta
BenchmarkRead-4                                      2622          3017          +15.06%
BenchmarkReadWithFieldsPerRecord-4                   2576          3041          +18.05%
BenchmarkReadWithoutFieldsPerRecord-4                2559          3086          +20.59%
BenchmarkReadLargeFields-4                           4666          5048          +8.19%
BenchmarkReadReuseRecord-4                           1430          1846          +29.09%
BenchmarkReadReuseRecordWithFieldsPerRecord-4        1420          1836          +29.30%
BenchmarkReadReuseRecordWithoutFieldsPerRecord-4     1455          1886          +29.62%
BenchmarkReadReuseRecordLargeFields-4                3399          3834          +12.80%

benchmark                                            old allocs     new allocs     delta
BenchmarkRead-4                                      16             16             +0.00%
BenchmarkReadWithFieldsPerRecord-4                   16             16             +0.00%
BenchmarkReadWithoutFieldsPerRecord-4                16             16             +0.00%
BenchmarkReadLargeFields-4                           24             24             +0.00%
BenchmarkReadReuseRecord-4                           6              6              +0.00%
BenchmarkReadReuseRecordWithFieldsPerRecord-4        6              6              +0.00%
BenchmarkReadReuseRecordWithoutFieldsPerRecord-4     6              6              +0.00%
BenchmarkReadReuseRecordLargeFields-4                12             12             +0.00%

benchmark                                            old bytes     new bytes     delta
BenchmarkRead-4                                      664           664           +0.00%
BenchmarkReadWithFieldsPerRecord-4                   664           664           +0.00%
BenchmarkReadWithoutFieldsPerRecord-4                664           664           +0.00%
BenchmarkReadLargeFields-4                           3936          3936          +0.00%
BenchmarkReadReuseRecord-4                           24            24            +0.00%
BenchmarkReadReuseRecordWithFieldsPerRecord-4        24            24            +0.00%
BenchmarkReadReuseRecordWithoutFieldsPerRecord-4     24            24            +0.00%
BenchmarkReadReuseRecordLargeFields-4                2976          2976          +0.00%
```

I didn't that because I didn't want to break the API of the original csv parser